requirements.txt
nextcord
mysql-connector-python